Workaround for docstring issue.
Particularly impactful when we index the docstring.
