+++ src/editor/codemirror/language-server/documentation.test.ts
@@ -3,31 +3,7 @@
  *
  * SPDX-License-Identifier: MIT
  */
-import {
-  renderDocumentation,
-  renderMarkdown,
-  splitDocString,
-} from "./documentation";
-
-describe("splitDocString", () => {
-  it("splits first para from remainder", () => {
-    const [first, remainder] = splitDocString(
-      "Frobs the wibble\n\nFurther information\n\nOn multiple lines\n"
-    );
-    expect(first).toEqual("Frobs the wibble");
-    expect(remainder).toEqual("Further information\n\nOn multiple lines\n");
-  });
-  it("returns undefined if no remainder", () => {
-    const [first, remainder] = splitDocString("Frobs the wibble");
-    expect(first).toEqual("Frobs the wibble");
-    expect(remainder).toBeUndefined();
-  });
-  it("returns something for empty string", () => {
-    const [first, remainder] = splitDocString("");
-    expect(first).toEqual("");
-    expect(remainder).toBeUndefined();
-  });
-});
+import { renderDocumentation, renderMarkdown } from "./documentation";
 
 describe("renderDocumentation", () => {
   it("sanitizes html", () => {
